Make zha isort-able.
